WIP:Preparing to migrate VUIDs
Initial changes to spec.py in order to parse json file of VUs
and convert string-based VUs into unique integer IDs that will be
used for the enum values.

Still need to get a clean json file and parse all of the VUs to
confirm that they're unique. Once that's done the actual conversion
will be a separate step.

Also probably want to move all of the mapping dicts and conversion
functions to their own file that can be shared with validation
codegen so that process for generating the IDs and for inserting
them into the generated code is exactly the same.