tests: Fix loader test validation
Fix a few validation errors in the loader test which showed up
when validation is enabled.

Thanks to Tony for catching these.

NOTE: If you have an Nvidia and Intel in the system and Optimus
is enabled, you will get errors during the vkEnumerateDeviceGroupsKHX
test which appear to be because of the way Optimus is calling down
with a non-NULL invalid structure for the groups structure.

Change-Id: I808ca01c719a28f55ef3581647f598a4bdd9d18e